+++ tests/testthat/test_updateMeta.R
@@ -50,3 +50,11 @@ test_that("Update Meta all_GADSdat", {
   expect_equal(changes_out$datList$df2, data.frame(df2$dat, v5 = c(1, 2)))
   expect_equal(changes_out$allLabels$varName, c("ID1", "v3", "ID1", "V2", "v5", "v5"))
 })
+
+test_that("illegal variable names", {
+  newDat_ill <- df1$dat
+  newDat_ill[, "Alter"] <- NA
+  mess <- capture_messages(out_both <- updateMeta(df1, newDat_ill))
+  expect_equal(mess[2], "Alter has been renamed to AlterVar\n")
+  expect_equal(names(out_both$dat), c("ID1", "V1", "AlterVar"))
+})